# Conflicts:
#	deploy/docs/Soil Water Characteristics-Equations.xls
#	deploy/installer/LANDIS-II-V7 PnET-Succession 3.5-b13-setup.exe
#	deploy/installer/PnET-Succession.txt
#	src/Cohort.cs
#	src/ISiteCohorts.cs
#	src/PlugIn.cs
#	src/PnET-Succession.csproj
#	src/SiteCohorts.cs

+++ src/ISiteCohorts.cs
@@ -10,7 +10,7 @@ namespace Landis.Extension.Succession.BiomassPnET
         float[] NetPsn { get; }
 
         float[] MaintResp{ get; }
-        
+
         float[] GrossPsn{ get; }
 
         float[] FolResp { get; }
@@ -54,7 +54,7 @@ namespace Landis.Extension.Succession.BiomassPnET
 
         int AgeMax { get; }
 
-        ushort WaterMax { get; }
+        float WaterMax { get; }
 
         uint BelowGroundBiomass { get; }
